requirements.txt
fastapi
uvicorn
httpx
pyyaml
pytest
streamlit
